refactor: move ExternalItemEditModal to food-item components
