+++ src/sections/item-group/components/ItemGroupEditModal.tsx
@@ -21,7 +21,7 @@ import {
 import { type TemplateItem } from '~/modules/diet/template-item/domain/templateItem'
 import { TemplateSearchModal } from '~/sections/search/components/TemplateSearchModal'
 import { ExternalTemplateSearchModal } from '~/sections/search/components/ExternalTemplateSearchModal'
-import { ExternalItemEditModal } from '~/sections/search/components/ExternalItemEditModal'
+import { ExternalItemEditModal } from '~/sections/food-item/components/ExternalItemEditModal'
 import { ItemEditModal } from '~/sections/food-item/components/ItemEditModal'
 import { RecipeIcon } from '~/sections/common/components/icons/RecipeIcon'
 import { RecipeEditModal } from '~/sections/recipe/components/RecipeEditModal'
